Completed exercise 02-flex-header
Added flex to both the header and the ul items so that all boxes are on one line like a webpage header should be. The left links and right links will always hug the edges no matter the page size. Tweaked some padding and margins so that all items are spaced nicely.