Completed exercise 03-flex-header-2
In the html file, split some items into a "left-side" container and a "right-side" container to allow an easier time for flexbox to move things around. Made it so the header container, ul container, and right-side, left-side were display: flex. Then simply added some gap and padding to make it look nice.